Fixes an oversight with the forked tongue (#94303)
## About The Pull Request

It was discovered that if a forked tongue is in an inventory slot, be it
your hands or pocket it gave you lizard speech. This is likely an
oversight that was missed for a long while. It seems likely it is
unintended behavior. This PR just adds a check to in the speechmod to
check if the items is a organ or not.
<details>
<summary>Screenshots</summary>

<img width="734" height="757" alt="image"
src="https://github.com/user-attachments/assets/f42d7993-0fad-4716-a18e-2db168253c98"
/>

<img width="470" height="631" alt="image"
src="https://github.com/user-attachments/assets/c0a4fe07-0e09-4aba-94e5-a62d5c1312ab"
/>

</details>

## Why It's Good For The Game

Less bugs and no silly behavior with a tongue not being inside of your
body.