Fixes runtimes from throwing fishing lines into water (#94502)
## About The Pull Request

We check reachability using an abstract /obj object for reaches over a
single tile, which caused a runtime as we tried to create an overlay for
a 0 width sprite. Probably shouldn't be making overlays for invisible
(due to lack of an icon) objects anyways.

## Changelog
:cl:
fix: Fixes runtimes from throwing fishing lines into water
/:cl: