Balance: Removes the override for the contractor baton (#6667)
## About The Pull Request

On the tin, this item's just too powerful and needs toned back down and
putting it back to TG default's a start.

## How This Contributes To The Nova Sector Roleplay Experience

Being able to permastun multiple secoffs is probably the most depressing
gameplay loop avaiable, this should hopefully be enough to discourage
that but we'll see.

## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
  
</details>

## Changelog
:cl:
balance: removed the Nova stamina damage Buff to the contractor baton
/:cl: