Removes the rest of the NRI mentions, outside of typepaths and the Blue Lizard wargame (#6654)
## About The Pull Request
Title.
No more (most of the) NRI mentions in-game.
It's all HC now, with the few exceptions that can't be simply stamped
out with a simple NRI -> HC swap.
## How This Contributes To The Nova Sector Roleplay Experience
Old lore is gone. Praise be the new lore.
## Proof of Testing
<details>
<summary>Screenshots/Videos</summary>
It's all text changes so I sincerely didn't bother; that choice beacon
wasn't used for anything but the NRI ERT that was previously removed, so
it won't affect anything either.
</details>

## Changelog
:cl: Stalkeros
del: Removes all the advanced choice beacon system's code that the
then-NRI had.
spellcheck: Removes a lot, or even most of the mentions of the NRI,
replaced with HC.
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>