Adds a fish trait extractor/injector to the marine utility node (and a bluespace fishing reel) (#94042)
## About The Pull Request
Huzzah! It's been a good while since I last added more fishing features
(instead I focused on small fixes now and then), but here I'm at it once
more.

This PR adds two designs to the Marine Utility techweb node (which
previously only had the bio-electric generator kit): The fish gene-gun
an as well as the bluespace fishing reel.

The fish gene-gun is a pretty easy concept. It's a device that takes
empty syringes. When used on a fish with an empty syringe loaded inside
it, It'll prompt the user to choose which trait to extract. Once the
trait is extracted, the syringe is replaced with a specialized 'fish
trait injector' containing said trait. You can then hit another fish
with the fish gene-gun again to inject the trait, or use it to remove
the injector, which can then be used on its own, or loaded into the
gene-gun later (the results are the same anyway). There are two caveats
though: Using the fish-gene gun on the same fish more than once
(extracting or injector) will kill it, and it also cannot be used on
dead fish. The fish can still be revived though.

The other thing is the bluespace fishing reel. It greatly increases the
cast range, and it also goes through walls and other solid objects and
is not disrupted by them. If used on objects, reeling the item in will
move it through wall and solid objects as well. Other than that. It's a
pretty standard fishing reel. However, it has one flaw, and that is it
cannot go through reinforced walls (as well as indestructible ones), so
no stealing from the armory etc.

## Why It's Good For The Game
I've been having some honest to god skill issues with aquariums at times
(they were also half-broken in the past but things have gone for the
better...). Some of the fish traits are good, some are bad, others are
utterly terrible (*looking at the anxious zip zap*) and kinda withhold
some of the potential stuff that can be done, which is what prompted me
to make this PR.

Also, bluespace fishing reel because ~~JoJo Vento Aureo reference~~ it's
such a neat concept.

## Changelog

:cl:
add: Two items have been added to the The Marine Utility tech node: a
fish-gene gun to add and remove traits to/from fish, and a bluespace
fishing reel.
/:cl:

---------

Co-authored-by: Jacquerel <hnevard@gmail.com>