You can soak a rag in a sink (#94501)
## About The Pull Request

You can soak a rag in a sink to fill it with water. 

Also while I was killing one letter variable I made it so that the
pickle jar interaction that sinks have puts the clean beaker back in
your hand.

## Why It's Good For The Game

A sink is the most reasonable place to soak a rag. I know y'all aren't
dumping beakers of water onto your rags at home.

## Changelog
:cl:

add: You can soak a rag in a sink

/:cl: