Fixes some rspack build errors (#94557)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
If you delete bun.lock for whatever reason, it'll update to a newer
rspack version which will break build 😃
So, I'm fixing it before we get hit by this. This comes bundled with
some other fixes from build time errors:

1. Rspack output path (as stated above)
2. Highlightjs doesn't bundle github dark anymore seemingly
3. Panel can read the wrong backend state, so a shim is inserted to
prevent it from erroneously suspending
4. Circular deps fixed with tgui-panel via
[plugin](https://rspack.rs/plugins/rspack/circular-dependency-rspack-plugin)
5. Fixed chat renderer functions via @ Kashargul
6. Coalesced chat settings imports courtesy @ Kashargul
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bug fixes before an issue is even made. The first will outright fail
builds
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->


<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Kashargul <144968721+Kashargul@users.noreply.github.com>