Fix bodythrow hitting yourself (#94553)
## About The Pull Request
Closes https://github.com/tgstation/tgstation/issues/59926

The issue lies with `+ world.tick_lag*2` blocking momentum. It's been
lying here for around 9 years. It only exists due to the way old way
keypress code used to work. Supposedly /tg/ used to move via keypress
repeats, meaning it used to spam the server more than once a tick. This
isn't a thing anymore, so this logic is now broken because the
prerequisites necessary are based on a now obsolete framework which used
to rely on duplicating keypress. Additionally see
https://github.com/tgstation/tgstation/issues/18670#issuecomment-292027037.

Incase anyone mentions that a thrownthing should preemptively be infront
of the the player's tile ahead of time (so, fire an extra tick for
leeway), the code already covers that. See `quickstart`.

## Why It's Good For The Game

Throwing a body won't hit yourself anymore, even if you're on meth.

## Changelog

:cl: ArchBTW
fix: Fix bodythrow hitting yourself
/:cl:

Co-authored-by: glue0000 <230859540+glue0000@users.noreply.github.com>