Excepts multiz pipes from mapping verb "check plumbing" (#94435)
## About The Pull Request

Multiz pipes will, on any map without 3 z levels, always be
"unconnected" to this mapping verb. This is because they're `TRINARY`,
and thus must fill `nodes` with three connection points in order to not
fail this verb, which they will never do unless they're on the center of
three z levels(and have pipes above and below them, which is the case
precisely nowhere). They need to keep `TRINARY` in order to keep
working, so I'm making an exception in the verb.

## Why It's Good For The Game

Check plumbing will no longer fill your chatbox with one million pipes
that are, in fact, connected.
Closes #94378

## Changelog
:cl:

No player facing changes. Or admin facing changes. This is just for us.

/:cl: