Adds the AI_COMPATIBLE flag to the Mauler exosuit (#94572)
## About The Pull Request

Just a very simple PR to fix an oversight where the Mauler exosuit
doesn't have the AI_COMPATIBLE flag, meaning as well as being unable to
be beaconed - which is intended - you also couldn't use an intellicard
to manually install an AI into it.
## Why It's Good For The Game

If a squad of nukies wants to shove an AI into their Mauler, then why
not? The base Marauder model which the Mauler is based on has AI
compatibility, so does the Seraph! Also brings parity with the Dark
Gygax which currently _can_ have an AI installed.
## Changelog
:cl:
fix: Fixed the inability to shove an intellicarded AI into a Mauler-type
exosuit.
/:cl: