Fixes problems with the giantslayer spear (#94620)
## About The Pull Request

Cleans up some of the code and grammar for the giantslayer spears.

Makes them not improvised, which is almost certainly an oversight given
they're not improvised weapons.

## Why It's Good For The Game

Code cleaning.

## Changelog
:cl:
code: Fixes up some of the code and grammar for the giantslayer spears.
fix: Giantslayer spears are not improvised and therefore do not break
when used to attack things.
/:cl: