Preserve inventory of mobs who turn into toilets (#94666)
## About The Pull Request

This PR changes it so that if you commit suicide via handheld experiment
scanner, all of your belongings are placed inside the resulting toilet's
cistern rather than just your brain.

## Why It's Good For The Game

It was pointed out that most forms of suicide that "destroy" the body
preserve the items, while this one does not.
This is largely a measure for ensuring that various items required for
traitor objectives or other reasons are not destroyed.

## Changelog

:cl:
fix: Meat toilets will now contain all of the belongings they were
wearing prior to transformation, which can be retrieved if the cistern
is open and the lid is closed.
/:cl: